Fix bug that re-installs ssh and breaks login

+++ cleanup.sh
@@ -1,4 +1,6 @@
 #! /bin/bash
 
 docker ps -aq | xargs docker rm -f
-docker images -aq| xargs docker rmi -f
+if [ "$1" == "remove" ]; then
+    docker images -aq | xargs docker rmi -f
+fi
